unicorn_contracts/pyproject.toml
[tool.poetry]
name = "contracts_service"
version = "0.2.0"
description = "Unicorn Properties Contact Service"
authors = ["Amazon Web Services"]
packages = [
    { include = "contracts_service", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.28.32"
aws-lambda-powertools = {extras = ["aws-sdk"], version = "^2.23.0"}
aws-xray-sdk = "^2.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
# pytest-cov = "^4.1.0"
# coverage = "^7.2.7"
requests = "^2.31.0"
moto = "^4.1.13"
importlib-metadata = "^6.8.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -vv -W ignore::UserWarning"
testpaths = [
    "./tests/unit",
    "./tests/integration",
]

[tool.ruff]
line-length = 150
